Add GitHub Actions for PRs
We need to run tests on all supported platforms for obvious reasons.
This initial configuration is based on an article from Alec Strong. It
is not sufficient because we need to test on multiple JDK versions and
we need to ensure that we are not running the same tests multiple times.
However, this is a good starting point.

https://www.alecstrong.com/2020/01/github-actions-mpp/